Restructure code to improve modularity
